requirements.txt
opencv-python~=4.5.4.60
fastapi~=0.70.1
starlette~=0.17.1
pandas~=1.3.5
numpy~=1.21.5
Pillow~=8.4.0